fix: Address PR review comments for canonical type mapping
- Change Decimal precision/scale from u8 to u16 to support PostgreSQL
  numerics which can exceed 255 digits
- Add explicit MySQL/MariaDB source handling in pools.rs to avoid
  incorrectly using MSSQL→MySQL mapper for MySQL sources
- Fix doc comment for PostgresFromCanonical (was incorrectly stating
  "PostgreSQL to canonical" instead of "Canonical to PostgreSQL")
- Improve MySQL boolean detection to handle tinyint(1) in column type
  string, since max_length from CHARACTER_MAXIMUM_LENGTH is 0 for
  numeric types
- Remove unreachable else branch in MysqlFromCanonical Varchar handling

+++ crates/mssql-pg-migrate/src/dialect/canonical.rs
@@ -56,10 +56,10 @@ pub enum CanonicalType {
     /// Exact decimal with precision and scale.
     /// Precision is total digits, scale is digits after decimal point.
     Decimal {
-        /// Total number of digits (1-65 for MySQL, 1-38 for MSSQL, unlimited for PG).
-        precision: u8,
+        /// Total number of digits (1-65 for MySQL, 1-38 for MSSQL, larger/unlimited for PG).
+        precision: u16,
         /// Number of digits after decimal point.
-        scale: u8,
+        scale: u16,
     },
     /// Money type with fixed precision (19,4).
     Money,
